Play around with how to handle workers
We could proxy tjrough the ZeroMQ messages to the workers, which would
solve a lot of issues, but also reduce our flexibility in future and
create new problems (e.g. if we want to send and heartbeat to the master
server, which one of the workers should do that?)

So, we'll likely route all communication through the scheduler thread
and see if that works out okay with high workloads.