Drop the C version of Spark
The point for C isn't really there when all you do with it is call
Python code anyway. Having the C code communicate with a lot of
Python/Perl/etc. workers just complicates the architecture of the tool.
So, we'll go with Python from here and see how far we get.